# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Soccer Connect API",
    "description": "Backend API for Soccer Connect - GTA soccer community platform",
    "version": "1.0.0",
    "contact": {
      "name": "Soccer Connect Team"
    }
  },
  "servers": [
    {
      "url": "https://soccer-connect-api.vercel.app/api",
      "description": "Production"
    },
    {
      "url": "http://localhost:8046/api",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
...
  "paths": {
...
    "/classifieds": {
      "get": {
        "tags": ["Classifieds"],
        "summary": "Get all classifieds with filters and pagination",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of classifieds" }
        }
      },
      "post": {
        "tags": ["Classifieds"],
        "summary": "Create a new classified",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "201": { "description": "Classified created" }
        }
      }
    },
    "/classifieds/user/my-classifieds": {
      "get": {
        "tags": ["Classifieds"],
        "summary": "Get classifieds created by current user",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "User's classifieds" }
        }
      }
    },
    "/classifieds/{id}": {
      "get": {
        "tags": ["Classifieds"],
        "summary": "Get single classified",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Classified details" }
        }
      },
      "put": {
        "tags": ["Classifieds"],
        "summary": "Update classified",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Classified updated" }
        }
      },
      "delete": {
        "tags": ["Classifieds"],
        "summary": "Delete classified",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Classified deleted" }
        }
      }
    },
    "/classifieds/{id}/respond": {
      "post": {
        "tags": ["Classifieds"],
        "summary": "Respond to a classified",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Response sent" }
        }
      }
    },
    "/classifieds/{id}/mark-filled": {
      "put": {
        "tags": ["Classifieds"],
        "summary": "Mark classified as filled",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Marked as filled" }
        }
      }
    },
...
    "/bookings": {
      "get": {
        "tags": ["Bookings"],
        "summary": "Get user's bookings",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "List of bookings" }
        }
      },
      "post": {
        "tags": ["Bookings"],
        "summary": "Create a new booking",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "201": { "description": "Booking created" }
        }
      }
    },
    "/bookings/{id}": {
      "get": {
        "tags": ["Bookings"],
        "summary": "Get single booking",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Booking details" }
        }
      },
      "put": {
        "tags": ["Bookings"],
        "summary": "Update booking",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Booking updated" }
        }
      },
      "delete": {
        "tags": ["Bookings"],
        "summary": "Delete a booking (only cancelled bookings)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Booking deleted" }
        }
      }
    },
    "/bookings/{id}/confirm": {
      "put": {
        "tags": ["Bookings"],
        "summary": "Confirm a booking",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Booking confirmed" }
        }
      }
    },
    "/bookings/{id}/cancel": {
      "put": {
        "tags": ["Bookings"],
        "summary": "Cancel a booking",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Booking cancelled" }
        }
      }
    },
...
  }
}